// src/api/generated/model/createOrdContractTemplateRequestDtoType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

/**
 * Тип договора
 */
export type CreateOrdContractTemplateRequestDtoType = typeof CreateOrdContractTemplateRequestDtoType[keyof typeof CreateOrdContractTemplateRequestDtoType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateOrdContractTemplateRequestDtoType = {
  service: 'service',
  mediation: 'mediation',
  additional: 'additional',
} as const;