// src/api/generated/model/ordIssuanceRuleTemplateDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { OrdIssuanceRuleTemplateDtoType } from './ordIssuanceRuleTemplateDtoType';
import type { OrdIssuanceRuleTemplateDtoDateStrategy } from './ordIssuanceRuleTemplateDtoDateStrategy';
import type { OrdIssuanceRuleTemplateDtoDateEndStrategy } from './ordIssuanceRuleTemplateDtoDateEndStrategy';
import type { OrdIssuanceRuleTemplateDtoActionType } from './ordIssuanceRuleTemplateDtoActionType';
import type { OrdIssuanceRuleTemplateDtoSubjectType } from './ordIssuanceRuleTemplateDtoSubjectType';
import type { OrdIssuanceRuleTemplateDtoFlagsItem } from './ordIssuanceRuleTemplateDtoFlagsItem';

export interface OrdIssuanceRuleTemplateDto {
  id: string;
  createdAt: string;
  updatedAt: string;
  name: string;
  /** ID мастер-комнаты (заказчик в выпускаемых договорах) */
  roomId: string;
  type: OrdIssuanceRuleTemplateDtoType;
  dateStrategy: OrdIssuanceRuleTemplateDtoDateStrategy;
  /** @nullable */
  fixedDate: string | null;
  dateEndStrategy: OrdIssuanceRuleTemplateDtoDateEndStrategy;
  /** @nullable */
  fixedDateEnd: string | null;
  /** @nullable */
  dateEndOffsetDays: number | null;
  /** @nullable */
  fixedAmount: string | null;
  autoGetCid: boolean;
  /** @nullable */
  actionType: OrdIssuanceRuleTemplateDtoActionType;
  /** @nullable */
  subjectType: OrdIssuanceRuleTemplateDtoSubjectType;
  flags: OrdIssuanceRuleTemplateDtoFlagsItem[];
}